*: remove EncodeResourceGroupTag
